# core/memory/working/projects/rate-my-set/IN/better-base-dev/openapi/generated/schema.yml
openapi: 3.0.3
info:
  title: Better Base API
  version: 0.1.0
  description: Documentation of API endpoints of Better Base
paths:
...
  /api/memberships/{id}/select:
    post:
      operationId: api_memberships_select_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Membership.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MembershipSelect'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MembershipSelect'
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipSelect'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSelect'
          description: ''
  /api/memberships/{id}/update-role:
    post:
      operationId: api_memberships_update_role_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Membership.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MembershipUpdateRole'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MembershipUpdateRole'
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipUpdateRole'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipUpdateRole'
          description: ''
...
  /api/users/{id}/check-delete:
    post:
      operationId: api_users_check_delete_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this User.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CheckUserDeletion'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CheckUserDeletion'
          application/json:
            schema:
              $ref: '#/components/schemas/CheckUserDeletion'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckUserDeletion'
          description: ''
...
  /api/users/{id}/update-uploaded-profile-image:
    post:
      operationId: api_users_update_uploaded_profile_image_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this User.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserUpdateUploadedProfileImage'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateUploadedProfileImage'
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateUploadedProfileImage'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdateUploadedProfileImage'
          description: ''
components:
  schemas:
...
    AccountReadOnly:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account_type:
          allOf:
          - $ref: '#/components/schemas/AccountTypeEnum'
          readOnly: true
        account_type_display:
          type: string
          readOnly: true
        name:
          type: string
          readOnly: true
        uploaded_profile_image:
          type: string
          format: uri
          readOnly: true
        fallback_name:
          type: string
          readOnly: true
        display_name:
          type: string
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - account_type
      - account_type_display
      - created
      - display_name
      - fallback_name
      - id
      - name
      - uploaded_profile_image
...
    AccountTypeEnum:
      enum:
      - personal
      - team
      type: string
      description: |-
        * `personal` - Personal
        * `team` - Team
...
    CheckUserDeletion:
      type: object
      properties:
        user:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        can_delete_user:
          type: boolean
          readOnly: true
        should_offer_manual_actions_before_deleting:
          type: boolean
          readOnly: true
        automated_actions_planned:
          type: object
          additionalProperties:
            enum:
            - delete-account
            - delete-membership
            type: string
          readOnly: true
        manual_actions_required:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items:
                enum:
                - transfer-ownership
                - delete-account
                - notify-other-owners
                type: string
          readOnly: true
        manual_actions_offered:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items:
                enum:
                - transfer-ownership
                - delete-account
                - notify-other-owners
                type: string
          readOnly: true
        account_ids_all_cleared:
          type: array
          items:
            type: integer
          readOnly: true
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/MembershipReadOnlyUserNotIncluded'
          readOnly: true
      required:
      - account_ids_all_cleared
      - automated_actions_planned
      - can_delete_user
      - manual_actions_offered
      - manual_actions_required
      - memberships
      - should_offer_manual_actions_before_deleting
      - user
...
    IsAuthenticatedEnum:
      enum:
      - true
      type: boolean
...
    MembershipReadOnlyUserNotIncluded:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          allOf:
          - $ref: '#/components/schemas/AccountReadOnly'
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        role_display:
          type: string
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - account
      - created
      - id
      - role
      - role_display
    MembershipSelect:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          allOf:
          - $ref: '#/components/schemas/AccountReadOnly'
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        role_display:
          type: string
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        last_selected_at:
          type: string
          format: date-time
          readOnly: true
          description: The last time the user created, switched to, or selected (etc.)
            this membership. This allows us to easily sort memberships by most recently
            selected.
      required:
      - account
      - created
      - id
      - last_selected_at
      - role
      - role_display
      - user
    MembershipUpdateRole:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          allOf:
          - $ref: '#/components/schemas/AccountReadOnly'
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        role:
          $ref: '#/components/schemas/RoleEnum'
        role_display:
          type: string
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - account
      - created
      - id
      - role
      - role_display
      - user
...
    RoleEnum:
      enum:
      - member
      - owner
      type: string
      description: |-
        * `member` - Member
        * `owner` - Owner
...
    UserReadOnly:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        email_is_verified:
          type: boolean
          readOnly: true
          title: Email Is Verified?
          description: Has this User's email address been verified?
        email_verified_as_of:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: When was this User's email address verified?
        name:
          type: string
          readOnly: true
        uploaded_profile_image:
          type: string
          format: uri
          readOnly: true
        is_authenticated:
          allOf:
          - $ref: '#/components/schemas/IsAuthenticatedEnum'
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
          title: Is Active?
          description: Designates whether this user should be treated as active. Potentially
            unselect this instead of deleting accounts.
        is_staff:
          type: boolean
          readOnly: true
          title: Is Staff?
          description: Designates whether the user can log into this admin site.
        is_superuser:
          type: boolean
          readOnly: true
          title: Superuser status
          description: Designates that this user has all permissions without explicitly
            assigning them.
        last_login:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - created
      - email
      - email_is_verified
      - email_verified_as_of
      - id
      - is_active
      - is_authenticated
      - is_staff
      - is_superuser
      - last_login
      - name
      - uploaded_profile_image
...
    UserUpdateUploadedProfileImage:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        email_is_verified:
          type: boolean
          readOnly: true
          title: Email Is Verified?
          description: Has this User's email address been verified?
        email_verified_as_of:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: When was this User's email address verified?
        name:
          type: string
          readOnly: true
        uploaded_profile_image:
          type: string
          format: uri
        is_authenticated:
          allOf:
          - $ref: '#/components/schemas/IsAuthenticatedEnum'
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
          title: Is Active?
          description: Designates whether this user should be treated as active. Potentially
            unselect this instead of deleting accounts.
        is_staff:
          type: boolean
          readOnly: true
          title: Is Staff?
          description: Designates whether the user can log into this admin site.
        is_superuser:
          type: boolean
          readOnly: true
          title: Superuser status
          description: Designates that this user has all permissions without explicitly
            assigning them.
        last_login:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - created
      - email
      - email_is_verified
      - email_verified_as_of
      - id
      - is_active
      - is_authenticated
      - is_staff
      - is_superuser
      - last_login
      - name
...
  securitySchemes:
    CsrfExemptSessionAuthentication:
      type: apiKey
      in: cookie
      name: sessionid
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid